pyproject.toml
[tool.poetry]
name = "fun-game"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
anthropic = "^0.37.1"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
openai = "^1.52.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"
pylint = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
